# Proto Model: removed some boost::accumulate uses