--- a/shared_model/backend/protobuf/commands/impl/proto_create_role.cpp
+++ b/shared_model/backend/protobuf/commands/impl/proto_create_role.cpp
@@ -5,22 +5,21 @@
 
 #include "backend/protobuf/commands/proto_create_role.hpp"
 
-#include <boost/range/numeric.hpp>
 #include "backend/protobuf/permissions.hpp"
 
 namespace shared_model {
   namespace proto {
 
     CreateRole::CreateRole(iroha::protocol::Command &command)
-        : create_role_{command.create_role()},
-          role_permissions_{boost::accumulate(
-              create_role_.permissions(),
-              interface::RolePermissionSet{},
-              [](auto &&acc, const auto &perm) {
-                acc.set(permissions::fromTransport(
-                    static_cast<iroha::protocol::RolePermission>(perm)));
-                return std::forward<decltype(acc)>(acc);
-              })} {}
+        : create_role_{command.create_role()}, role_permissions_{[&command] {
+            auto &perms_in = command.create_role().permissions();
+            interface::RolePermissionSet perms_out;
+            for (const auto &perm : perms_in) {
+              perms_out.set(permissions::fromTransport(
+                  static_cast<iroha::protocol::RolePermission>(perm)));
+            }
+            return perms_out;
+          }()} {}
 
     const interface::types::RoleIdType &CreateRole::roleName() const {
       return create_role_.role_name();