--- a/shared_model/backend/protobuf/queries/impl/proto_get_transactions.cpp
+++ b/shared_model/backend/protobuf/queries/impl/proto_get_transactions.cpp
@@ -5,20 +5,18 @@
 
 #include "backend/protobuf/queries/proto_get_transactions.hpp"
 
-#include <boost/range/numeric.hpp>
+#include <boost/range/adaptor/transformed.hpp>
 
 namespace shared_model {
   namespace proto {
 
     GetTransactions::GetTransactions(iroha::protocol::Query &query)
         : get_transactions_{query.payload().get_transactions()},
-          transaction_hashes_{boost::accumulate(
-              get_transactions_.tx_hashes(),
-              TransactionHashesType{},
-              [](auto &&acc, const auto &hash) {
-                acc.push_back(crypto::Hash::fromHexString(hash));
-                return std::forward<decltype(acc)>(acc);
-              })} {}
+          transaction_hashes_{boost::copy_range<TransactionHashesType>(
+              get_transactions_.tx_hashes()
+              | boost::adaptors::transformed([](const auto &hash) {
+                  return crypto::Hash::fromHexString(hash);
+                }))} {}
 
     const GetTransactions::TransactionHashesType &
     GetTransactions::transactionHashes() const {