--- a/shared_model/backend/protobuf/query_responses/impl/proto_role_permissions_response.cpp
+++ b/shared_model/backend/protobuf/query_responses/impl/proto_role_permissions_response.cpp
@@ -5,7 +5,6 @@
 
 #include "backend/protobuf/query_responses/proto_role_permissions_response.hpp"
 
-#include <boost/range/numeric.hpp>
 #include "backend/protobuf/permissions.hpp"
 #include "utils/string_builder.hpp"
 
@@ -16,14 +15,16 @@ namespace shared_model {
         iroha::protocol::QueryResponse &query_response)
         : role_permissions_response_{query_response
                                          .role_permissions_response()},
-          role_permissions_{boost::accumulate(
-              role_permissions_response_.permissions(),
-              interface::RolePermissionSet{},
-              [](auto &&permissions, const auto &permission) {
-                permissions.set(permissions::fromTransport(
-                    static_cast<iroha::protocol::RolePermission>(permission)));
-                return std::forward<decltype(permissions)>(permissions);
-              })} {}
+          role_permissions_{[&query_response] {
+            auto &perms_in =
+                query_response.role_permissions_response().permissions();
+            interface::RolePermissionSet perms_out;
+            for (const auto &perm : perms_in) {
+              perms_out.set(permissions::fromTransport(
+                  static_cast<iroha::protocol::RolePermission>(perm)));
+            }
+            return perms_out;
+          }()} {}
 
     const interface::RolePermissionSet &
     RolePermissionsResponse::rolePermissions() const {