--- a/shared_model/backend/protobuf/query_responses/impl/proto_roles_response.cpp
+++ b/shared_model/backend/protobuf/query_responses/impl/proto_roles_response.cpp
@@ -5,19 +5,14 @@
 
 #include "backend/protobuf/query_responses/proto_roles_response.hpp"
 
-#include <boost/range/numeric.hpp>
+#include <boost/range/iterator_range_core.hpp>
 
 namespace shared_model {
   namespace proto {
 
     RolesResponse::RolesResponse(iroha::protocol::QueryResponse &query_response)
         : roles_response_{query_response.roles_response()},
-          roles_{boost::accumulate(roles_response_.roles(),
-                                   RolesIdType{},
-                                   [](auto &&roles, const auto &role) {
-                                     roles.emplace_back(role);
-                                     return std::move(roles);
-                                   })} {}
+          roles_{boost::copy_range<RolesIdType>(roles_response_.roles())} {}
 
     const RolesResponse::RolesIdType &RolesResponse::roles() const {
       return roles_;