Signed-off-by: Mikhail Boldyrev <miboldyrev@gmail.com>

[IR-651](https://jira.hyperledger.org/browse/IR-651)

<!-- You will not see HTML commented line in Pull Request body -->
<!-- Optional sections may be omitted. Just remove them or write None -->

<!-- ### Requirements -->
<!-- * Filling out the template is required. Any pull request that does not include enough information to be reviewed in a timely manner may be closed at the maintainers' discretion. -->
<!-- * All new code must have code coverage above 70% (https://docs.codecov.io/docs/about-code-coverage). -->
<!-- * CircleCI builds must be passed. -->
<!-- * Critical and blocker issues reported by Sorabot must be fixed. -->
<!-- * Branch must be rebased onto base branch (https://soramitsu.atlassian.net/wiki/spaces/IS/pages/11173889/Rebase+and+merge+guide). -->


### Description of the Change

Reworked code to avoid `boost::accumulate`.

<!-- We must be able to understand the design of your change from this description. If we can't get a good idea of what the code will be doing from the description here, the pull request may be closed at the maintainers' discretion. -->
<!-- Keep in mind that the maintainer reviewing this PR may not be familiar with or have worked with the code here recently, so please walk us through the concepts. -->

### Benefits

Over 100 times faster objects accumulation. Hoperfully fixes some fuzzing time-exceeded issues.

<!-- What benefits will be realized by the code change? -->

### Possible Drawbacks 

<!-- What are the possible side-effects or negative impacts of the code change? -->
<!-- If no drawbacks, explicitly mention this (write None) -->

### Usage Examples or Tests *[optional]*

<!-- Point reviewers to the test, code example or documentation which shows usage example of this feature -->


before:
```
time ./builds/fuzzing/test_bin/find_fuzz /tmp/clusterfuzz-testcase-find_fuzz-5719861837168640
Reading 23853 bytes from /tmp/clusterfuzz-testcase-find_fuzz-5719861837168640
Execution successful
./builds/fuzzing/test_bin/find_fuzz   9.98s user 0.12s system 99% cpu 10.130 total

time ./builds/fuzzing/test_bin/torii_fuzz /tmp/clusterfuzz-testcase-torii_fuzz-5753054091018240
Reading 36605 bytes from /tmp/clusterfuzz-testcase-torii_fuzz-5753054091018240
Execution successful
./builds/fuzzing/test_bin/torii_fuzz   8.41s user 0.49s system 99% cpu 8.901 total
```
after:
```
time ./builds/fuzzing/test_bin/find_fuzz /tmp/clusterfuzz-testcase-find_fuzz-5719861837168640
Reading 23853 bytes from /tmp/clusterfuzz-testcase-find_fuzz-5719861837168640
Execution successful
./builds/fuzzing/test_bin/find_fuzz   0.06s user 0.02s system 99% cpu 0.079 total

time ./builds/fuzzing/test_bin/torii_fuzz /tmp/clusterfuzz-testcase-torii_fuzz-5753054091018240
Reading 36605 bytes from /tmp/clusterfuzz-testcase-torii_fuzz-5753054091018240
Execution successful
./builds/fuzzing/test_bin/torii_fuzz   0.07s user 0.00s system 99% cpu 0.077 total
```


### Alternate Designs *[optional]*

<!-- Explain what other alternates were considered and why the proposed version was selected -->